requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
boto3==1.28.74
botocore==1.31.74
Brotli==1.1.0
bs4==0.0.1
calmjs.parse==1.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clr-loader==0.2.6
colorama==0.4.6
crispy-bootstrap5==0.7
cron-descriptor==1.4.0
csscompressor==0.9.5
cssselect2==0.7.0
Django==4.2.5
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-compressor==4.4
django-cors-headers==4.3.0
django-crispy-forms==2.0
django-dotenv==1.4.2
django-redis-cache==3.0.1
django-storages==1.14.2
django-timezone-field==6.0.1
django-utils-six==2.0
dnspython==2.4.2
eventlet==0.33.3
fonttools==4.43.1
frozenlist==1.4.0
gevent==23.9.1
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
hiredis==2.2.3
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
jmespath==1.0.1
joblib==1.3.2
kombu==5.3.2
multidict==6.0.4
nltk==3.8.1
numpy==1.26.0
outcome==1.2.0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
ply==3.11
prompt-toolkit==3.0.39
psycopg2==2.9.9
pycparser==2.21
pydyf==0.8.0
pyee==8.2.2
pyphen==0.14.0
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pythonnet==3.0.3
pytz==2023.3.post1
rcssmin==1.1.1
redis==3.5.3
regex==2023.10.3
requests==2.31.0
rjsmin==1.2.1
s3transfer==0.7.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stripe==7.0.0
textblob==0.17.1
tinycss2==1.2.1
to==0.3
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
vine==5.0.0
wcwidth==0.2.8
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==10.4
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
zopfli==0.2.3
